+++ tests/infrastructures/Kubernetes/Transcriber/ReplicationControllerTranscriberTest.php
@@ -26,6 +26,7 @@ declare(strict_types=1);
 namespace Teknoo\Tests\East\Paas\Infrastructures\Kubernetes\Transcriber;
 
 use Maclof\Kubernetes\Client as KubeClient;
+use Maclof\Kubernetes\Models\ReplicationController;
 use Maclof\Kubernetes\Repositories\ReplicationControllerRepository;
 use PHPUnit\Framework\TestCase;
 use Teknoo\Recipe\Promise\PromiseInterface;
@@ -115,16 +116,26 @@ class ReplicationControllerTranscriberTest extends TestCase
                 ['replicationControllers', [], $srRepo],
             ]);
 
-        $srRepo->expects(self::exactly(2))
-            ->method('exists')
-            ->willReturnOnConsecutiveCalls(false, true);
+        $srRepo->expects(self::any())
+            ->method('setLabelSelector')
+            ->willReturnSelf();
 
-        $srRepo->expects(self::once())
+        $srRepo->expects(self::exactly(2))
+            ->method('first')
+            ->willReturnOnConsecutiveCalls(
+                null,
+                new ReplicationController(['metadata' => ['name' => 'foo']]),
+            );
+
+        $srRepo->expects(self::exactly(2))
             ->method('create')
             ->willReturn(['foo']);
 
+        $srRepo->expects(self::never())
+            ->method('update');
+
         $srRepo->expects(self::once())
-            ->method('update')
+            ->method('delete')
             ->willReturn(['foo']);
 
         $promise = $this->createMock(PromiseInterface::class);
@@ -202,21 +213,98 @@ class ReplicationControllerTranscriberTest extends TestCase
                 ['replicationControllers', [], $srRepo],
             ]);
 
-        $srRepo->expects(self::exactly(2))
-            ->method('exists')
-            ->willReturnOnConsecutiveCalls(false, true);
+        $srRepo->expects(self::any())
+            ->method('setLabelSelector')
+            ->willReturnSelf();
 
-        $srRepo->expects(self::once())
+        $srRepo->expects(self::exactly(2))
+            ->method('first')
+            ->willReturnOnConsecutiveCalls(
+                null,
+                new ReplicationController(['metadata' => ['name' => 'foo']]),
+            );
+
+        $srRepo->expects(self::exactly(2))
             ->method('create')
             ->willReturn(['foo']);
 
+        $srRepo->expects(self::never())
+            ->method('update');
+
         $srRepo->expects(self::once())
-            ->method('update')
+            ->method('delete')
             ->willReturn(['foo']);
 
         $promise = $this->createMock(PromiseInterface::class);
         $promise->expects(self::exactly(2))->method('success')->with(['foo']);
         $promise->expects(self::never())->method('fail');
+
+        self::assertInstanceOf(
+            ReplicationControllerTranscriber::class,
+            $this->buildTranscriber()->transcribe($cd, $kubeClient, $promise)
+        );
+    }
+
+    public function testErrorOnFetching()
+    {
+        $kubeClient = $this->createMock(KubeClient::class);
+        $cd = $this->createMock(CompiledDeploymentInterface::class);
+
+        $cd->expects(self::once())
+            ->method('foreachPod')
+            ->willReturnCallback(function (callable $callback) use ($cd) {
+                $image1 = new Image('foo', '/foo', true, '7.4', ['foo' => 'bar']);
+                $image1 = $image1->withRegistry('repository.teknoo.run');
+                $image2 = new Image('bar', '/bar', true, '7.4', []);
+                $image2 = $image2->withRegistry('repository.teknoo.run');
+
+                $volume1 = new Volume('foo1', ['foo' => 'bar'], '/foo', '/mount');
+                $volume2 = new Volume('bar1', ['bar' => 'foo'], '/bar', '/mount');
+
+                $c1 = new Container('c1', 'foo', '7.4', [80], ['foo' => $volume1->import('/foo')], []);
+                $c2 = new Container('c2', 'bar', '7.4', [80], ['bar' => $volume2->import('/bar')], []);
+
+                $pod1 = new Pod('p1', 1, [$c1]);
+                $pod2 = new Pod('p2', 1, [$c2]);
+
+                $callback($pod1, ['foo' => ['7.4' => $image1]], ['foo' => $volume1], 'default_namespace');
+                $callback($pod2, ['bar' => ['7.4' => $image2]], ['bar' => $volume2], 'default_namespace');
+                return $cd;
+            });
+
+        $repo = $this->createMock(ReplicationControllerRepository::class);
+        $kubeClient->expects(self::any())
+            ->method('__call')
+            ->with('replicationControllers')
+            ->willReturn($repo);
+
+        $repo->expects(self::any())
+            ->method('setLabelSelector')
+            ->willReturnSelf();
+
+        $call = 0;
+        $repo->expects(self::exactly(2))
+            ->method('first')
+            ->willReturnCallback(
+                function () use (&$call) {
+                    if (0 < $call++) {
+                        throw new \DomainException('foo');
+                    }
+
+                    return null;
+                },
+            );
+
+        $repo->expects(self::once())
+            ->method('create')
+            ->willReturn(['foo']);
+
+        $repo->expects(self::never())
+            ->method('update');
+
+        $promise = $this->createMock(PromiseInterface::class);
+        $promise->expects(self::once())->method('success')->with(['foo']);
+        $promise->expects(self::once())->method('fail');
 
         self::assertInstanceOf(
             ReplicationControllerTranscriber::class,
@@ -257,17 +345,32 @@ class ReplicationControllerTranscriberTest extends TestCase
             ->with('replicationControllers')
             ->willReturn($repo);
 
+        $repo->expects(self::any())
+            ->method('setLabelSelector')
+            ->willReturnSelf();
+
         $repo->expects(self::exactly(2))
-            ->method('exists')
-            ->willReturnOnConsecutiveCalls(false, true);
+            ->method('first')
+            ->willReturnOnConsecutiveCalls(
+                null,
+                new ReplicationController(['metadata' => ['name' => 'foo']]),
+            );
 
-        $repo->expects(self::once())
+        $call = 0;
+        $repo->expects(self::exactly(2))
             ->method('create')
-            ->willReturn(['foo']);
+            ->willReturnCallback(
+                function () use (&$call) {
+                    if (0 < $call++) {
+                        throw new \DomainException('foo');
+                    }
 
-        $repo->expects(self::once())
-            ->method('update')
-            ->willThrowException(new \Exception());
+                    return ['foo'];
+                },
+            );
+
+        $repo->expects(self::never())
+            ->method('update');
 
         $promise = $this->createMock(PromiseInterface::class);
         $promise->expects(self::once())->method('success')->with(['foo']);